# gen/proto/swagger/vince/api/v1/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "vince/api/v1/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Vince"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/event": {
      "post": {
        "operationId": "Vince_SendEvent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "n",
            "description": "/ EventName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "url",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "d",
            "description": "Domain",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "w",
            "description": "Screen width",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "h",
            "description": "Hash mode",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "ip",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "ua",
            "description": "user agent",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "r",
            "description": "Referrer",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Vince"
        ]
      }
    },
    "/v1/cluster": {
      "get": {
        "operationId": "Vince_GetCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetClusterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "Vince"
        ]
      }
    },
    "/v1/cluster/apply": {
      "post": {
        "operationId": "Vince_ApplyCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ApplyClusterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "config.name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "config.nodes",
            "description": "This is a request variable of the map type. The query format is \"map_name[key]=value\", e.g. If the map name is Age, the key type is string, and the value type is integer, the query parameter is expressed as Age[\"bob\"]=18",
            "in": "query",
            "required": false
          }
        ],
        "tags": [
          "Vince"
        ]
      }
    },
    "/v1/login": {
      "post": {
        "operationId": "Vince_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "publicKey",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          },
          {
            "name": "generate",
            "description": "When true, the token will be generated by the server.",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "ttl",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Vince"
        ]
      }
    },
    "/v1/version": {
      "get": {
        "operationId": "Vince_Version",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Build"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "Vince"
        ]
      }
    }
  },
  "definitions": {
    "ClientAuth": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "token": {
          "type": "string"
        },
        "serverId": {
          "type": "string"
        }
      }
    },
    "ConfigNode": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/ClientAuth"
        },
        "bootstrap": {
          "type": "boolean",
          "description": "True if this node is used to bootstrap the cluster."
        },
        "address": {
          "type": "string"
        }
      }
    },
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "v1ApplyClusterResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "string"
        }
      }
    },
    "v1Build": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        }
      }
    },
    "v1ClusterConfig": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "nodes": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/ConfigNode"
          }
        }
      }
    },
    "v1GetClusterResponse": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/v1ClusterConfig"
        }
      }
    },
    "v1LoginResponse": {
      "type": "object",
      "properties": {
        "auth": {
          "$ref": "#/definitions/ClientAuth"
        }
      }
    }
  }
}